move config to subdirectory
